refactor: 기존 테스트 리팩토링
